# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  student: 'student',
  teacher: 'teacher',
} as const;

export interface AuthUser {
  id: string;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface SignupRequest {
  email: string;
  /**
     * @minLength 8
     * @maxLength 256
     */
  password: string;
  /** @minLength 1 */
  name: string;
  role?: UserRole;
}

export interface LoginRequest {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface Class {
  id: string;
  teacherId: string;
  name: string;
  joinCode: string;
  createdAt: string;
}

export interface CreateClassRequest {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
}

export interface Assignment {
  id: string;
  classId: string;
  title: string;
  instructions: string;
  dueAt?: string | null;
  createdAt: string;
}

export interface CreateAssignmentRequest {
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /** @maxLength 5000 */
  instructions?: string;
  dueAt?: string | null;
}

export interface ProblemStep {
  /** @minLength 1 */
  prompt: string;
  /** @minLength 1 */
  answer: string;
}

export interface Problem {
  id: string;
  assignmentId: string;
  position: number;
  prompt: string;
  steps: ProblemStep[];
  createdAt: string;
}

export interface CreateProblemRequest {
  /**
     * @minLength 1
     * @maxLength 5000
     */
  prompt: string;
  /** @minItems 1 */
  steps: ProblemStep[];
}

export type AssignmentWithProblems = Assignment & {
  problems: Problem[];
};