# lib/api-zod/src/generated/types/assignment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Assignment {
  id: string;
  classId: string;
  title: string;
  instructions: string;
  dueAt?: Date | null;
  createdAt: Date;
}