# lib/api-zod/src/generated/types/assignmentWithProblems.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Assignment } from './assignment';
import type { Problem } from './problem';

export type AssignmentWithProblems = Assignment & {
  problems: Problem[];
};