# lib/api-zod/src/generated/types/class.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Class {
  id: string;
  teacherId: string;
  name: string;
  joinCode: string;
  createdAt: Date;
}