# lib/api-zod/src/generated/types/createAssignmentRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAssignmentRequest {
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /** @maxLength 5000 */
  instructions?: string;
  dueAt?: Date | null;
}