# lib/api-zod/src/generated/types/createClassRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateClassRequest {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
}