# lib/api-zod/src/generated/types/createProblemRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemStep } from './problemStep';

export interface CreateProblemRequest {
  /**
     * @minLength 1
     * @maxLength 5000
     */
  prompt: string;
  /** @minItems 1 */
  steps: ProblemStep[];
}