# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemStep } from './problemStep';

export interface Problem {
  id: string;
  assignmentId: string;
  position: number;
  prompt: string;
  steps: ProblemStep[];
  createdAt: Date;
}