# lib/api-zod/src/generated/types/problemStep.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProblemStep {
  /** @minLength 1 */
  prompt: string;
  /** @minLength 1 */
  answer: string;
}